Initial Set Up
Initial Set Up for a discord bot that will take in a screenshot from a text channel, check if the user has already acquired the item in the screen shot, and then attribute points to them if they have not already acquired the item before. Can pull up a point leaderboard for current season.